Adapted EBCOperation TLS to single threaded runtime
So far only 1 EBC was allowed to run at a time on a runtime thread. This
did not work for sync depth first scheduling of message processing. In
sync mode all EBC will be run on the same thread. Pushing EBC
continuations on a stack in TLS solved the issue.